# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisSummary": {
        "description": "An analysis without its report -- what a list endpoint returns.",
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionProgress"
            },
            "title": "Criteria",
            "type": "array"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "stage": {
            "default": "",
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          },
          "totals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisTotals"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "analysis_id",
          "document_id",
          "status"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
...
      "AnalyzeRequest": {
        "properties": {
          "criteria": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Criterion ids to run. Omit for all five.",
            "title": "Criteria"
          },
          "document_id": {
            "description": "The document to analyse. Required: there is no implicit current document.",
            "title": "Document Id",
            "type": "integer"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "Body_upload_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "The contract, as a PDF.",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_documents_post",
        "type": "object"
      },
...
      "DocumentOut": {
        "properties": {
          "chunks": {
            "default": 0,
            "title": "Chunks",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "ingested_at": {
            "default": "",
            "title": "Ingested At",
            "type": "string"
          },
          "last_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastAnalysisOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "spine_source": {
            "default": "none",
            "title": "Spine Source",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "filename"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "Error": {
        "description": "The body of every 4xx and 5xx this API returns.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nCompliance analysis of PDF contracts, and cited chat over them.\n\nUpload a contract to get a `document_id`, then bind everything else to it:\nanalysis and chat both require one, and retrieval is scoped to it in the\nlibrary rather than in this API, so an answer cannot cite another contract.\n\nAnalyses are jobs. `POST /analyses` returns an `analysis_id` in under a second\nand the run takes a minute or more; poll `GET /analyses/{analysis_id}` or\nsubscribe to its `/events`. Chat streams by default and can return one JSON\nbody instead.\n\nErrors are always `{\"error\": {\"code\", \"message\", \"hint\"}}`. `code` is stable\nand `hint` says what to do next, which is what a model driving this API through\nan MCP server needs in order to recover.\n",
    "title": "Contract Analyzer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyses": {
      "get": {
        "description": "Live and stored analyses merged, so a run from a previous boot or from `make analyze` is listed too. `document_id` is required: a global list is a KPI question, and `/metrics/runs` is where it belongs.",
        "operationId": "index_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response Index Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One document's analyses, newest first",
        "tags": [
          "analyses"
        ]
      },
      "post": {
        "description": "Returns immediately with an `analysis_id`; poll `GET /analyses/{id}` or subscribe to `GET /analyses/{id}/events`. A submission that matches one already queued or running returns that analysis with `200` instead of starting a second run -- send an `Idempotency-Key` header to force a new one.",
        "operationId": "submit_analyses_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Queue an analysis of one contract",
        "tags": [
          "analyses"
        ]
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "description": "Answered from the live job when this process is running it, and from the stored record otherwise -- a finished analysis returns its report after a restart. `detail=summary` omits quotes and rationale -- the MCP default, because a full report is a lot of context to put in a model's window.",
        "operationId": "detail_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "detail",
            "required": false,
            "schema": {
              "default": "full",
              "enum": [
                "full",
                "summary"
              ],
              "title": "Detail",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "An analysis, with its report once it is done",
        "tags": [
          "analyses"
        ]
      }
    },
...
    "/documents": {
      "get": {
        "description": "What a client binds a session to, with enough to draw a row for each.\n\n`last_analysis` comes from one query over the whole list rather than one\nper document: this is the endpoint a library table renders from, and a\nStreamlit script re-runs it on every click.",
        "operationId": "index_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response Index Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Every stored contract, newest first",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Parses, chunks and embeds the PDF, and returns the `document_id` every other endpoint needs. Each upload mints a new id even for identical bytes, so two sessions working on the same contract cannot see each other's analyses.",
        "operationId": "upload_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing or wrong X-API-Key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such document or analysis."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The resource is busy, or already in the requested state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The upload is over api_max_upload_mb."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Only PDF uploads are supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request body did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unhandled server error; the X-Trace-Id is on every log line."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An upstream call failed after its retries."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A key or a dependency the operation needs is not configured."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Upload a contract and index it",
        "tags": [
          "documents"
        ]
      }
    },
...
  },
...
}